Support contextvars on python>=3.7
